projeto finalizado

+++ Aliquota/Aliquota.Domain.Services/Entities/ContaEntity.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int Numero { get; set; }
-        public int Saldo { get; set; }
-        public int ValorAplicado { get; set; }
+        public double Saldo { get; set; }
+        public double ValorAplicado { get; set; }
     }
 }
